Add content to recipe pages: image, description, ingredients, steps
This commit adds the necessary content to the recipe pages for avocado-toast, bbq-chicken-pizza, and spicy-tuna-roles. The goal is to provide a more complete and informative recipe experience for users.

Previously, the recipe pages only contained a title (h1 heading). This commit enriches the pages by adding an image of the dish, a description heading with a paragraph, an ingredients heading with an unordered list, and a steps heading with an ordered list. This offers users a more comprehensive view of each recipe, making it easier for them to follow and execute.

Additional details:

Edited the URLs on the parent index.html file to relative links

Edited the following pages:
- avocado-toast/avocado-toast.html
- bbq-chicken-pizza/bbq-chicken-pizza.html
- spicy-tuna-roles/spicy-tuna-roles.html

Example of added content includes:
- An image of the dish
- Description heading and paragraph
- Ingredients heading and unordered list
- Steps heading and ordered list

modified:   index.html
modified:   recipes/avocado-toast/avocado-toast.html
modified:   recipes/bbq-chicken-pizza/bbq-chicken-pizza.html
modified:   recipes/spicy-tuna-roles/spicy-tuna-roles.html